\name{housingexp}
\docType{data}
\alias{housingexp}
\title{Boston Housing Data with 3-Way Interactions}
\description{
	This data set is developed based on the original Boston Housing Data. We use \code{MEDV} as the repsonse \code{y}. In the matrix of variables \code{x}, we include all the other 13 variables, their second and third orders, their two-way and three-way interactions.
}
\usage{data(housingexp)}
\source{\url{http://stat.wharton.upenn.edu/~buja/STAT-541/boston.dat}}
\keyword{datasets}